hz: improve precision of rate limiting
By estimating the minimum sleep time clock_nanosleep() will use up, we can
determine whether or not sleeping will take us past our desired quantum. If we
are too close to the quantum for a clock_nanosleep() to work well for us, then
we can just spin reading the CPU clock until time runs out.

For the most part, this should still be efficient, because it will sleep for
the majority of the spare time.

Signed-off-by: Steven Noonan <steven@uplinklabs.net>